swami_sales/requirements.txt

coreapi==2.3.3
coreschema==0.0.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0


